+++ account_invoice_export/__manifest__.py
@@ -1,11 +1,11 @@
 
 {
-    'name': 'Account Invoice in Turkey currency ',
-    'version': '8.0.2',
-    'website': 'https://www.odoo.com',
-    'author': 'Kirankumar',
+    'name': 'Account Invoice in Turkish Lira',
+    'version': '12.0.0.1',
+    'website': 'https://altinkaya.com.tr',
+    'author': 'Yiğit Budak, Kirankumar',
     'category': 'Invoice',
-    'summary': 'This Module is used to calculate total invoice amount in turkey currency',
+    'summary': 'This Module is used to calculate total invoice amount in Turkish Lira',
     'description': """This Module is use to Export Invoice""",
     'depends': ['base','account'],
     'installable' : True,
