+++ altinkaya_sales/views/product_views.xml
@@ -83,9 +83,19 @@
 			</xpath>
 
 			<xpath expr="//group[@name='properties']" position="before">
+				<group name="fields">
+				</group>
+			</xpath>
+
+			<xpath expr="//group[@name='fields']" position="inside">
     			<xpath expr="//group[2]/label[1]" position="move"/>
 				<xpath expr="//group[2]/div[1]" position="move"/>
 			</xpath>
+
+			<xpath expr="//group[@name='fields']" position="inside">
+				<xpath expr="//field[@name='hs_code_id']" position="move"/>
+			</xpath>
+
 
            <!-- <xpath expr="//field[@name='description']" position="after">
                 <group name="Old_price" string="Old Prices">
